split mappeR_util with numba
